OCT-111: moved DeleteCatalogQuery
