+++ components/catalogs/back/src/Infrastructure/Persistence/Catalog/DeleteCatalogQuery.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Infrastructure\Persistence\Catalog;
 
-use Akeneo\Catalogs\Application\Persistence\DeleteCatalogQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Catalog\DeleteCatalogQueryInterface;
 use Doctrine\DBAL\Connection;
 use Ramsey\Uuid\Uuid;
